requirements.txt
astroid==2.3.3
autopep8==1.5
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-cors-headers==3.1.1
djangorestframework==3.9.2
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.2
pycodestyle==2.5.0
pylint==2.4.4
python-decouple==3.3
pytz==2019.1
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
typed-ast==1.4.1
whitenoise==4.1.4
wrapt==1.11.2
